refactor(evaluation): delete deprecated pre-Stickler comparator engine
Evaluation runs entirely through Stickler (service.py), so the dead
pre-Stickler engine is removed:

- Delete evaluation/service_legacy.py (old custom EvaluationService; zero
  live callers) and evaluation/comparator.py (Comparator classes +
  compare_exact/fuzzy/numeric/hungarian/semantic/values/date helpers)
- Relocate the two functions still used at runtime:
  - compare_llm -> llm_comparator.py (its only caller, the Stickler-
    registered LLMComparator)
  - fuzz_score + strip_punctuation_space -> new evaluation/text_matching.py
    (used by assessment/ocr_grounding for OCR value matching)
- evaluation/__init__.py no longer re-exports the compare_* helpers
- Add "DateComparator": "Date" to service._normalize_comparator_name so DATE
  fields render with a clean label in evaluation reports (live-path gap)
- Remove 13 skipped tests in test_evaluation_service.py that mocked the
  removed compare_values/_count_classifications internals; delete
  test_date_comparator.py (covered the deleted legacy helpers). Live-path
  DATE coverage remains in test_stickler_mapper.py::TestDateMethodMapping
- Retire evaluation_methods_demo.ipynb (demoed the removed compare_values API)
- Docs + CHANGELOG

No YAML-config or runtime-behavior impact; affects only custom code importing
those internal symbols directly. Verified: evaluation suite (69) + config
validation (96) + llm_comparator/ocr_grounding/assessment (184) green; ruff
and basedpyright clean; import sanity confirms relocated symbols resolve and
deleted modules are gone.
