+++ CHANGELOG.md
@@ -84,6 +84,8 @@ are migrated automatically on read — no manual edit is required.** See the
 
 ### Removed
 
+- **Deprecated pre-Stickler evaluation engine deleted.** Now that evaluation runs entirely through the Stickler library (`evaluation/service.py`), the dead legacy modules have been removed: `evaluation/service_legacy.py` (the old custom `EvaluationService`, no live callers) and `evaluation/comparator.py` (the `Comparator` classes and `compare_exact`/`compare_fuzzy`/`compare_numeric`/`compare_hungarian`/`compare_semantic`/`compare_values`/`compare_date` helpers). The two functions still needed at runtime were relocated: `compare_llm` now lives in `evaluation/llm_comparator.py` (its only caller, the Stickler-registered `LLMComparator`), and `fuzz_score`/`strip_punctuation_space` moved to a new dependency-free `evaluation/text_matching.py` (used by OCR bounding-box grounding). The `idp_common.evaluation` package no longer re-exports the `compare_*` helpers, and the `evaluation_methods_demo.ipynb` notebook (which demoed the removed `compare_values` API) is retired. No YAML-config or runtime-behavior impact; affects only custom code that imported those internal symbols directly.
+
 - **Granular assessment is retired and its service deleted.** The `GranularAssessmentService`, the `granular_service.py` module, and the `extraction.confidence.granular` config field no longer exist. Large-list assessment is now handled by **standalone large-list batching** (`extraction.confidence.list_batch_size`, default 25): the model scores each large list in sequential batches, then any dropped rows are re-scored by a bounded missing-row retry, so per-cell coverage reaches 100%. On a 120-row bank statement this was **~78% cheaper** than granular with equal accuracy and full per-cell confidence + geometry. **Migration is a no-op:** leftover `granular.*` keys still validate and are ignored. The standalone Assessment step's `postAssessment` pipeline-hook point is also removed. See [Granular Assessment Retirement](docs/migration-granular-retirement.md).
 
 - **`idp_common` code-level breaking changes (affects custom scripts and notebooks, not YAML config).** The config auto-migration keeps *runtime* behavior working, but code touching the removed Python symbols breaks:
