+++ lib/idp_common_pkg/idp_common/evaluation/__init__.py
@@ -8,16 +8,6 @@ This module provides services and models for evaluating document extraction resu
 using the Stickler library for structured object comparison.
 """
 
-# Legacy comparator functions (deprecated - kept for backward compatibility)
-from idp_common.evaluation.comparator import (
-    compare_date,
-    compare_exact,
-    compare_fuzzy,
-    compare_hungarian,
-    compare_numeric,
-    compare_values,
-)
-
 # Stickler integration components
 from idp_common.evaluation.llm_comparator import LLMComparator
 
@@ -31,7 +21,7 @@ from idp_common.evaluation.models import (
     SectionEvaluationResult,
 )
 
-# Stickler-based evaluation service (replaces legacy implementation)
+# Stickler-based evaluation service
 from idp_common.evaluation.service import EvaluationService
 from idp_common.evaluation.stickler_mapper import SticklerConfigMapper
 
@@ -42,18 +32,11 @@ __all__ = [
     "AttributeEvaluationResult",
     "SectionEvaluationResult",
     "DocumentEvaluationResult",
-    # Main service (now Stickler-based)
+    # Main service (Stickler-based)
     "EvaluationService",
     # Stickler components
     "SticklerConfigMapper",
     "LLMComparator",
     # Metrics
     "calculate_metrics",
-    # Legacy comparison functions (deprecated)
-    "compare_values",
-    "compare_exact",
-    "compare_numeric",
-    "compare_fuzzy",
-    "compare_hungarian",
-    "compare_date",
 ]
