+++ lib/idp_common_pkg/tests/unit/evaluation/test_evaluation_service.py
@@ -28,7 +28,6 @@ import pytest
 # Finally import application modules
 from idp_common.evaluation.service import EvaluationService
 from idp_common.evaluation.models import (
-    EvaluationMethod,
     AttributeEvaluationResult,
     SectionEvaluationResult,
 )
@@ -154,475 +153,6 @@ class TestEvaluationService:
         )
 
         return doc
-
-    @pytest.mark.skip(
-        reason="Internal method removed in Stickler migration - see test_evaluation_service_stickler.py"
-    )
-    def test_init(self, mock_config):
-        """Test initialization with configuration."""
-        service = EvaluationService(
-            region="us-west-2", config=mock_config, max_workers=5
-        )
-
-        assert service.region == "us-west-2"
-        assert service.max_workers == 5
-        # Note: These attributes don't exist in the Stickler-based implementation
-        # assert service.default_model == "anthropic.claude-3-sonnet-20240229-v1:0"
-        # assert service.default_temperature == 0.0
-        # assert service.default_top_k == 5
-        # assert "You are an evaluator" in service.default_system_prompt
-        # assert "Compare" in service.default_task_prompt
-
-    @pytest.mark.skip(reason="Internal method removed in Stickler migration")
-    def test_get_attributes_for_class(self, service):
-        """Test getting attributes for a document class."""
-        # Test with existing class
-        invoice_attrs = service._get_attributes_for_class("invoice")
-        assert len(invoice_attrs) == 3
-        assert invoice_attrs[0].name == "invoice_number"
-        assert invoice_attrs[0].evaluation_method == EvaluationMethod.EXACT
-        assert invoice_attrs[1].name == "invoice_date"
-        assert invoice_attrs[1].evaluation_method == EvaluationMethod.FUZZY
-        assert invoice_attrs[1].evaluation_threshold == 0.9
-        assert invoice_attrs[2].name == "total_amount"
-        assert invoice_attrs[2].evaluation_method == EvaluationMethod.LLM
-
-        # Test with non-existent class
-        unknown_attrs = service._get_attributes_for_class("unknown")
-        assert len(unknown_attrs) == 0
-
-        # Test case insensitivity
-        invoice_attrs_upper = service._get_attributes_for_class("INVOICE")
-        assert len(invoice_attrs_upper) == 3
-
-    @pytest.mark.skip(reason="Internal method removed in Stickler migration")
-    @patch("idp_common.s3.get_json_content")
-    def test_load_extraction_results(self, mock_get_json_content, service):
-        """Test loading extraction results from S3."""
-        # Test with inference_result wrapper
-        mock_get_json_content.return_value = {
-            "inference_result": {
-                "invoice_number": "INV-123",
-                "invoice_date": "2023-05-08",
-            }
-        }
-
-        extraction_results, confidence_scores = service._load_extraction_results(
-            "s3://bucket/path"
-        )
-        assert extraction_results == {
-            "invoice_number": "INV-123",
-            "invoice_date": "2023-05-08",
-        }
-        assert confidence_scores == {}
-
-        # Test without wrapper
-        mock_get_json_content.return_value = {
-            "invoice_number": "INV-123",
-            "invoice_date": "2023-05-08",
-        }
-
-        extraction_results, confidence_scores = service._load_extraction_results(
-            "s3://bucket/path"
-        )
-        assert extraction_results == {
-            "invoice_number": "INV-123",
-            "invoice_date": "2023-05-08",
-        }
-        assert confidence_scores == {}
-
-        # Test with error
-        mock_get_json_content.side_effect = Exception("S3 error")
-        extraction_results, confidence_scores = service._load_extraction_results(
-            "s3://bucket/path"
-        )
-        assert extraction_results == {}
-        assert confidence_scores == {}
-
-    @pytest.mark.skip(reason="Internal method removed in Stickler migration")
-    def test_count_classifications_both_empty(self, service):
-        """Test counting classifications when both values are empty."""
-        with patch(
-            "idp_common.evaluation.comparator.compare_values"
-        ) as mock_compare_values:
-            tn, fp, fn, tp, fp1, fp2, score, reason = service._count_classifications(
-                attr_name="test_attr",
-                expected=None,
-                actual=None,
-                evaluation_method=EvaluationMethod.EXACT,
-                threshold=0.8,
-            )
-
-            assert tn == 1
-            assert fp == 0
-            assert fn == 0
-            assert tp == 0
-            assert fp1 == 0
-            assert fp2 == 0
-            assert score == 1.0
-            assert "missing" in reason.lower()
-
-            # Mock compare_values was not called
-            mock_compare_values.assert_not_called()
-
-    @pytest.mark.skip(reason="Internal method removed in Stickler migration")
-    def test_count_classifications_expected_empty_actual_not(self, service):
-        """Test counting classifications when expected is empty but actual is not."""
-        with patch(
-            "idp_common.evaluation.comparator.compare_values"
-        ) as mock_compare_values:
-            tn, fp, fn, tp, fp1, fp2, score, reason = service._count_classifications(
-                attr_name="test_attr",
-                expected=None,
-                actual="value",
-                evaluation_method=EvaluationMethod.EXACT,
-                threshold=0.8,
-            )
-
-            assert tn == 0
-            assert fp == 1
-            assert fn == 0
-            assert tp == 0
-            assert fp1 == 1
-            assert fp2 == 0
-            assert score == 0.0
-            assert reason is None
-
-            # Mock compare_values was not called
-            mock_compare_values.assert_not_called()
-
-    @pytest.mark.skip(reason="Internal method removed in Stickler migration")
-    def test_count_classifications_expected_not_empty_actual_empty(self, service):
-        """Test counting classifications when expected is not empty but actual is."""
-        with patch(
-            "idp_common.evaluation.comparator.compare_values"
-        ) as mock_compare_values:
-            tn, fp, fn, tp, fp1, fp2, score, reason = service._count_classifications(
-                attr_name="test_attr",
-                expected="value",
-                actual=None,
-                evaluation_method=EvaluationMethod.EXACT,
-                threshold=0.8,
-            )
-
-            assert tn == 0
-            assert fp == 0
-            assert fn == 1
-            assert tp == 0
-            assert fp1 == 0
-            assert fp2 == 0
-            assert score == 0.0
-            assert reason is None
-
-            # Mock compare_values was not called
-            mock_compare_values.assert_not_called()
-
-    @pytest.mark.skip(reason="Internal method removed in Stickler migration")
-    def test_count_classifications_both_not_empty_match(self, service):
-        """Test counting classifications when both values are not empty and match."""
-        with patch(
-            "idp_common.evaluation.service.compare_values"
-        ) as mock_compare_values:
-            # Configure mock to return a match
-            mock_compare_values.return_value = (True, 1.0, "Values match")
-
-            tn, fp, fn, tp, fp1, fp2, score, reason = service._count_classifications(
-                attr_name="test_attr",
-                expected="value1",
-                actual="value1",
-                evaluation_method=EvaluationMethod.EXACT,
-                threshold=0.8,
-            )
-
-            assert tn == 0
-            assert fp == 0
-            assert fn == 0
-            assert tp == 1
-            assert fp1 == 0
-            assert fp2 == 0
-            assert score == 1.0
-            # The reason might be None in the actual implementation
-            # We're just checking that compare_values was called
-            mock_compare_values.assert_called_once()
-
-    @pytest.mark.skip(reason="Internal method removed in Stickler migration")
-    def test_count_classifications_both_not_empty_no_match(self, service):
-        """Test counting classifications when both values are not empty but don't match."""
-        with patch(
-            "idp_common.evaluation.service.compare_values"
-        ) as mock_compare_values:
-            # Configure mock to return no match
-            mock_compare_values.return_value = (False, 0.0, "Values don't match")
-
-            tn, fp, fn, tp, fp1, fp2, score, reason = service._count_classifications(
-                attr_name="test_attr",
-                expected="value1",
-                actual="value2",
-                evaluation_method=EvaluationMethod.EXACT,
-                threshold=0.8,
-            )
-
-            assert tn == 0
-            assert fp == 1
-            assert fn == 0
-            assert tp == 0
-            assert fp1 == 0
-            assert fp2 == 1
-            assert score == 0.0
-            # The reason might be None in the actual implementation
-            # We're just checking that compare_values was called
-            mock_compare_values.assert_called_once()
-
-    @pytest.mark.skip(reason="Internal method removed in Stickler migration")
-    def test_evaluate_single_attribute_match(self, service):
-        """Test evaluating a single attribute that matches."""
-        with patch(
-            "idp_common.evaluation.comparator.compare_values"
-        ) as mock_compare_values:
-            # Configure mock to return a match
-            mock_compare_values.return_value = (True, 1.0, "Values match")
-
-            with patch.object(service, "_count_classifications") as mock_count:
-                # Configure mock to return appropriate values
-                mock_count.return_value = (0, 0, 0, 1, 0, 0, 1.0, "Values match")
-
-                result, metrics = service._evaluate_single_attribute(
-                    attr_name="invoice_number",
-                    expected_value="INV-123",
-                    actual_value="INV-123",
-                    evaluation_method=EvaluationMethod.EXACT,
-                    evaluation_threshold=0.8,
-                    document_class="invoice",
-                    attr_description="The invoice number",
-                )
-
-                # Check result
-                assert result.name == "invoice_number"
-                assert result.expected == "INV-123"
-                assert result.actual == "INV-123"
-                assert result.matched is True
-                assert result.score == 1.0
-                assert result.reason == "Values match"
-                assert result.evaluation_method == "EXACT"
-                assert result.evaluation_threshold is None  # Not included for EXACT
-
-                # Check metrics
-                assert metrics["tp"] == 1
-                assert metrics["fp"] == 0
-                assert metrics["fn"] == 0
-                assert metrics["tn"] == 0
-                assert metrics["fp1"] == 0
-                assert metrics["fp2"] == 0
-
-    @pytest.mark.skip(reason="Internal method removed in Stickler migration")
-    def test_evaluate_single_attribute_no_match(self, service):
-        """Test evaluating a single attribute that doesn't match."""
-        with patch(
-            "idp_common.evaluation.comparator.compare_values"
-        ) as mock_compare_values:
-            # Configure mock to return no match
-            mock_compare_values.return_value = (False, 0.0, "Values don't match")
-
-            with patch.object(service, "_count_classifications") as mock_count:
-                # Configure mock to return appropriate values
-                mock_count.return_value = (0, 1, 0, 0, 0, 1, 0.0, "Values don't match")
-
-                result, metrics = service._evaluate_single_attribute(
-                    attr_name="invoice_number",
-                    expected_value="INV-123",
-                    actual_value="INV-456",
-                    evaluation_method=EvaluationMethod.EXACT,
-                    evaluation_threshold=0.8,
-                    document_class="invoice",
-                    attr_description="The invoice number",
-                )
-
-                # Check result
-                assert result.name == "invoice_number"
-                assert result.expected == "INV-123"
-                assert result.actual == "INV-456"
-                assert result.matched is False
-                assert result.score == 0.0
-                assert result.reason == "Values don't match"
-                assert result.evaluation_method == "EXACT"
-                assert result.evaluation_threshold is None  # Not included for EXACT
-
-                # Check metrics
-                assert metrics["tp"] == 0
-                assert metrics["fp"] == 1
-                assert metrics["fn"] == 0
-                assert metrics["tn"] == 0
-                assert metrics["fp1"] == 0
-                assert metrics["fp2"] == 1
-
-    @pytest.mark.skip(reason="Internal method removed in Stickler migration")
-    def test_evaluate_single_attribute_fuzzy(self, service):
-        """Test evaluating a single attribute with fuzzy matching."""
-        with patch(
-            "idp_common.evaluation.comparator.compare_values"
-        ) as mock_compare_values:
-            # Configure mock to return a match
-            mock_compare_values.return_value = (
-                True,
-                0.92,
-                "Values match with fuzzy comparison",
-            )
-
-            with patch.object(service, "_count_classifications") as mock_count:
-                # Configure mock to return appropriate values
-                mock_count.return_value = (
-                    0,
-                    0,
-                    0,
-                    1,
-                    0,
-                    0,
-                    0.92,
-                    "Values match with fuzzy comparison",
-                )
-
-                result, metrics = service._evaluate_single_attribute(
-                    attr_name="invoice_date",
-                    expected_value="2023-05-08",
-                    actual_value="May 8, 2023",
-                    evaluation_method=EvaluationMethod.FUZZY,
-                    evaluation_threshold=0.9,
-                    document_class="invoice",
-                    attr_description="The invoice date",
-                )
-
-                # Check result
-                assert result.name == "invoice_date"
-                assert result.expected == "2023-05-08"
-                assert result.actual == "May 8, 2023"
-                assert result.matched is True
-                assert result.score == 0.92
-                assert result.reason == "Values match with fuzzy comparison"
-                assert result.evaluation_method == "FUZZY"
-                assert result.evaluation_threshold == 0.9  # Included for FUZZY
-
-    @pytest.mark.skip(reason="Internal method removed in Stickler migration")
-    def test_evaluate_single_attribute_unconfigured(self, service):
-        """Test evaluating an unconfigured attribute."""
-        with patch(
-            "idp_common.evaluation.comparator.compare_values"
-        ) as mock_compare_values:
-            # Configure mock to return a match
-            mock_compare_values.return_value = (True, 0.95, "Values match")
-
-            with patch.object(service, "_count_classifications") as mock_count:
-                # Configure mock to return appropriate values
-                mock_count.return_value = (0, 0, 0, 1, 0, 0, 0.95, "Values match")
-
-                result, metrics = service._evaluate_single_attribute(
-                    attr_name="unconfigured_attr",
-                    expected_value="value1",
-                    actual_value="value1",
-                    evaluation_method=EvaluationMethod.LLM,
-                    evaluation_threshold=0.8,
-                    document_class="invoice",
-                    attr_description="Unconfigured attribute",
-                    is_unconfigured=True,
-                )
-
-                # Check result
-                assert result.name == "unconfigured_attr"
-                assert result.matched is True
-                assert "Default method" in result.reason
-
-                # Check metrics
-                assert metrics["tp"] == 1
-                assert metrics["fp"] == 0
-
-    @pytest.mark.skip(reason="Internal method removed in Stickler migration")
-    @patch("idp_common.s3.get_json_content")
-    def test_evaluate_section(self, mock_get_json_content, service):
-        """Test evaluating a document section."""
-        # Create a section
-        section = Section(section_id="1", classification="invoice", page_ids=["1", "2"])
-
-        # Define expected and actual results
-        expected_results = {
-            "invoice_number": "INV-123",
-            "invoice_date": "2023-05-08",
-            "total_amount": "$100.00",
-        }
-
-        actual_results = {
-            "invoice_number": "INV-123",
-            "invoice_date": "May 8, 2023",
-            "total_amount": "$100.00",
-        }
-
-        # Mock the _evaluate_single_attribute method
-        with patch.object(service, "_evaluate_single_attribute") as mock_evaluate:
-            # Configure mock to return successful matches for all attributes
-            mock_evaluate.side_effect = [
-                (
-                    AttributeEvaluationResult(
-                        name="invoice_number",
-                        expected="INV-123",
-                        actual="INV-123",
-                        matched=True,
-                        score=1.0,
-                        reason="Exact match",
-                        evaluation_method="EXACT",
-                    ),
-                    {"tp": 1, "fp": 0, "fn": 0, "tn": 0, "fp1": 0, "fp2": 0},
-                ),
-                (
-                    AttributeEvaluationResult(
-                        name="invoice_date",
-                        expected="2023-05-08",
-                        actual="May 8, 2023",
-                        matched=True,
-                        score=0.92,
-                        reason="Fuzzy match",
-                        evaluation_method="FUZZY",
-                        evaluation_threshold=0.9,
-                    ),
-                    {"tp": 1, "fp": 0, "fn": 0, "tn": 0, "fp1": 0, "fp2": 0},
-                ),
-                (
-                    AttributeEvaluationResult(
-                        name="total_amount",
-                        expected="$100.00",
-                        actual="$100.00",
-                        matched=True,
-                        score=1.0,
-                        reason="Exact match",
-                        evaluation_method="LLM",
-                    ),
-                    {"tp": 1, "fp": 0, "fn": 0, "tn": 0, "fp1": 0, "fp2": 0},
-                ),
-            ]
-
-            # Patch the calculate_metrics function
-            with patch(
-                "idp_common.evaluation.metrics.calculate_metrics"
-            ) as mock_metrics:
-                mock_metrics.return_value = {
-                    "precision": 1.0,
-                    "recall": 1.0,
-                    "f1_score": 1.0,
-                }
-
-                # Evaluate section
-                result = service.evaluate_section(
-                    section=section,
-                    expected_results=expected_results,
-                    actual_results=actual_results,
-                )
-
-                # Check result
-                assert result.section_id == "1"
-                assert result.document_class == "invoice"
-                assert len(result.attributes) == 3
-
-                # Check metrics
-                assert result.metrics["precision"] == 1.0
-                assert result.metrics["recall"] == 1.0
-                assert result.metrics["f1_score"] == 1.0
 
     @patch("idp_common.s3.get_json_content")
     @patch("idp_common.evaluation.service.EvaluationService._process_section")
